??Replaced type int to double

+++ Examples/Еxample004_Div/Program.cs
@@ -1,4 +1,4 @@
 ﻿// See https://aka.ms/new-console-template for more information
-int numberA = 10;
-int numberB = 5;
+double numberA = 10;
+double numberB = 5;
 Console.WriteLine(numberA / numberB);
